+++ ootrando_overworldmap_hamsda/changelog.md
@@ -11,6 +11,7 @@
 - changed layouts to accomodate smaller resolutions better
   - moved pinned locations between items and map in horizontal layout
   - added scrollbar to pinned locations in horizontal layout
+- added `Plant Magic Beans` setting and removed bean planting `user_setting`
 
 ## 3.7.3.0